Extract naming functions to typescript/name.ts
Move all name-generation functions (module names, file names, field
names, profile class names, method names, etc.) from utils.ts and
profile.ts into a dedicated name.ts file.
